move grpc pool manager
