+++ node/worker/main.py
@@ -6,7 +6,7 @@ import os
 import asyncio
 import traceback
 import resource
-from node.grpc_pool_manager import get_grpc_pool_instance, close_grpc_pool
+from node.server.grpc_pool_manager import get_grpc_pool_instance, close_grpc_pool
 import psutil
 
 logger = get_logger(__name__)
